feat: share action card component between operation viewer and editor
